2004-08-06  Jay Rossiter <cryptographite@users.sf.net>  1.4.5-cvs-b95
        * Fix: Prevent foreach() error from being displayed when previewing
          watermark previews.

        * Fix: Logging into Gallery as a non-admin and then trying to reset
          the admin password failed.  The logged in user was used and the
          resetadmin file was ignored


git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery@7048 57fcd75e-5312-0410-8df3-f5eb6fbb1595

+++ classes/AlbumItem.php
@@ -475,8 +475,11 @@ class AlbumItem {
 			return (0);
 		}
                 $name = $this->image->name;
-		foreach (glob($dir . "/$name.preview*.$type") as $oldpreview) {
-			unlink($oldpreview);
+		$oldpreviews = glob($dir . "/$name.preview*.$type");
+		if (!empty($oldpreviews) && is_array($oldpreviews)) {
+			foreach ($oldpreviews as $oldpreview) {
+				unlink($oldpreview);
+			}
 		}
 		if ($preview) {
 			$previewtag = "preview" . time();
